requirement.txt
newspaper3k
streamlit
sumy
